+++ src/test/php/unittest/tests/Baseless.class.php
@@ -1,0 +1,5 @@
+<?php namespace unittest\tests;
+
+class Baseless {
+  
+}